separate qa and fix dependencies. clean setup/dockerfile
