--- qa-requirements.txt
@@ -1,2 +1,0 @@
-flask==3.0.0
-gunicorn==21.2.0
