--- radifox/qa/templates/subject.html
@@ -1,77 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
-          integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-    <style>
-        html {
-            font-size: 0.5rem;
-        }
-
-        @media (min-width: 576px) {
-            html {
-                font-size: 0.625rem;
-            }
-        }
-
-        @media (min-width: 768px) {
-            html {
-                font-size: 0.75rem;
-            }
-        }
-
-        @media (min-width: 992px) {
-            html {
-                font-size: 0.875rem;
-            }
-        }
-
-        @media (min-width: 1200px) {
-            html {
-                font-size: 1rem;
-            }
-        }
-    </style>
-    <title>RADIFOX QA - {{ subject_id }}</title>
-</head>
-<body>
-<nav class="navbar navbar-expand-md sticky-top navbar-dark bg-dark">
-    <a class="navbar-brand" href="{{ url_for('index') }}">RADIFOX QA</a>
-    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon"></span>
-    </button>
-    <div class="collapse navbar-collapse" id="navbarSupportedContent">
-    <ul class="navbar-nav mr-auto">
-      <li class="nav-item {% if not session['qa_mode'] or session['qa_mode'] == 'conversion' %}active{% endif %}">
-        <a class="nav-link" href="{{ url_for('set_mode', mode='conversion') }}">Conversion</a>
-      </li>
-      <li class="nav-item {% if session['qa_mode'] == 'processing' %}active{% endif %}">
-        <a class="nav-link" href="{{ url_for('set_mode', mode='processing') }}">Processing</a>
-      </li>
-    </ul>
-    <span class="navbar-text">
-        <a href="{{ url_for('project', project_id=project_id) }}/">{{ project_id | upper() }}</a>&nbsp;/&nbsp;{{ subject_id }}
-    </span>
-  </div>
-</nav>
-<div class="container-fluid">
-    <br/>
-    <h1 class="text-center">Sessions</h1>
-    <hr/>
-    {% for session_id in sessions %}
-    <div class="row">
-        <div class="col">
-            <a href="{{ url_for('qa_page', project_id=project_id, subject_id=subject_id, session_id=session_id) }}">
-                <h3 class="text-center">{{ session_id }}</h3>
-            </a>
-        </div>
-    </div>
-    {% endfor %}
-</div>
-</body>
-</html>
