--- requirements.txt
@@ -1,7 +1,0 @@
-nibabel==5.1.0
-numpy==1.26.0
-pillow==10.0.1
-pydicom==2.4.3
-pyyaml==6.0.1
-scipy==1.11.3
-radifox-utils==1.0.2
